// agricultural-pesticide-platform/types/entities.ts
export interface Category {
  category_id: string
  category_name: string
  description: string
  category_img?: string // Category image URL
  is_active: boolean
  created_at: Date
  updated_at: Date
}